[Base] add exception creation callback & error print on except (#701)
This pr add two functions in shambase/exception.hpp

```c++
    void set_exception_gen_callback(void (*callback)(std::string msg));
    void exception_gen_callback(std::string msg);
```

the first one set a callback that is called on exception to print an
error in the terminal.
The second one is called whenever make_except_with_loc is called.

This ensure that we get error messages when an exception is created,
typically if a chain of exception is created this ensure that we get all
of them instead of the last one.

This pr also add a new env variable `SHAMLOG_ERR_ON_EXCEPT` that enable
exception reporting to err only if set to `1`. If not set it will
default to `1`.