[MHD] Factorize the MHD solver + fix a bug in the debug dump format. (#800)
This has been tested on the sound wave test. The debug format works as
is (dumping 8 extra fields).
The div cleaning field is still exploding due to the conservation term
in its evolution equation, thus the default MHD is set to not include
divergence cleaning. I will try a different cleaning method whilst I
investigate why this is happening.

---------

Co-authored-by: David--Cléris Timothée <timothee.davidcleris@proton.me>