[Env] Allow installing shamrock executable with pip (#912)
# Allow installing shamrock executable with pip

This PR finally implement a way to install both the pylib and the
executable when using `pip install` (thanks to the solution provided in
https://github.com/pypa/setuptools/issues/210#issuecomment-449696043).

Additionally `_pysetup.py` was moved to env helpers.

Lastly, this PR removes `--pylib` and will systematically copy
`setup.py` to the builddir if the env supports it.