[CMake] add check on git submodule commit hash (#925)
# Checking that git submodule are on cloned & on correct commit

Most peoples use `git pull` or `git checkout` to update their local
repo. However, this does not update/clone submodules. As a result every
time I update a submodule this results in breakage on local repos
because everyone forget about this.

So this PR add a check to verify that each submodule is cloned and at
the correct commit and prompt the correct command to fix it if not.