[CMake] Allow disabling SYCL header inclusion in cmake (#933)
Adding SYCL headers as include flags is necessary only when using
Shamrock in an IDE with LSP or clang-tidy for them to find SYCL headers.
On a supercomputer or in a packaged version this is not required.

This PR add the `SHAMROCK_ADD_SYCL_INCLUDES` cmake option to toggle them
(on by default).