[Sys] add micro GPU benchmark to post startup benchmarks (#932)
# Expand shamrock tiny benchmarks to GPU ones

When using `--benchmark-mpi` (yes it should be renamed), new GPU
benchmarks were added. Namely, sappy (bandwidth) and multiply additions
rotations to compute flops.

On a RTXA5000 the benchmarks report:
```
Running micro benchamrks : 
 - p2p bandwith    : 9.1475e+08 B.s^-1 (ranks : 0 -> 0) (loops : 111) 
 - saxpy (f32_4)   : 7.108e+11 B.s^-1 (min = 7.1e+11, max = 7.1e+11, avg = 7.1e+11) (9.1e+00 ms) 
 - add_mul (f32_4) : 1.482e+13 flops (min = 1.5e+13, max = 1.5e+13, avg = 1.5e+13) (1.7e+01 ms) 
 - add_mul (f64_4) : 2.042e+11 flops (min = 2.0e+11, max = 2.0e+11, avg = 2.0e+11) (1.2e+03 ms) 
```

Where the capability of a RTXA5000 are:
```
 - saxpy (f32_4)   : 7.68e+11 B.s^-1
 - add_mul (f32_4) : 2.777e+13 flops
 - add_mul (f64_4) : 4.33e+11 flops
```
So bandwidth is close but flops are a factor 2 below (which is kind of
expected as we don't use the other type of computing units, e.g. tensor,
ray, texture, ...)