[Algs] add binned_average and binned_computation for histogram calculations (#1152)
binned_average does basically the same thing as binned_sum but decides
by the counter.
binned_computation takes as a parameter a function and similarly
conducts a computation. I created it because I had type issues
(inferring the Tret type) when I tried to directly use binned_compute.
binned_computation is for instance used to compute the surface density
of a disc.

---------

Co-authored-by: David--Cléris Timothée <timothee.davidcleris@proton.me>